feat: update init grade manager
