+++ saas/backend/api/management/views/__init__.py
@@ -9,11 +9,7 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 from .application import ManagementGroupApplicationViewSet
-from .grade_manager import (
-    ManagementGradeManagerInitViewSet,
-    ManagementGradeManagerMemberViewSet,
-    ManagementGradeManagerViewSet,
-)
+from .grade_manager import ManagementGradeManagerMemberViewSet, ManagementGradeManagerViewSet
 from .group import (
     ManagementGradeManagerGroupViewSet,
     ManagementGroupActionPolicyViewSet,
@@ -26,7 +22,6 @@ from .user import ManagementUserGradeManagerGroupViewSet, ManagementUserGradeMan
 __all__ = [
     "ManagementGradeManagerMemberViewSet",
     "ManagementGradeManagerViewSet",
-    "ManagementGradeManagerInitViewSet",
     "ManagementGradeManagerGroupViewSet",
     "ManagementGroupViewSet",
     "ManagementGroupMemberViewSet",
