+++ saas/backend/trans/open_management.py
@@ -14,16 +14,9 @@ from typing import Dict, List
 from pydantic.tools import parse_obj_as
 
 from backend.biz.action import ActionBiz
-from backend.biz.group import GroupTemplateGrantBean
-from backend.biz.policy import PolicyBean, PolicyBeanList
+from backend.biz.policy import PolicyBeanList
 from backend.biz.role import RoleBiz, RoleInfoBean
-from backend.service.constants import ADMIN_USER
-from backend.service.models.policy import ResourceGroupList
-from backend.service.models.subject import Subject
-from backend.service.role import AuthScopeAction, AuthScopeSystem
-from backend.util.uuid import gen_uuid
 
-from .constants import ManagementCommonActionNameEnum, ManagementGroupNameSuffixEnum
 from .open import OpenCommonTrans, OpenPolicy
 
 
@@ -182,123 +175,3 @@ class GradeManagerTrans(ManagementCommonTrans):
             # 替换掉data里原有的authorization_scopes
             data["authorization_scopes"] = self._preprocess_authorization_scopes(data["authorization_scopes"])
         return RoleInfoBean.from_partial_data(data)
-
-    def init_role_info(self, data):
-        """
-        创建初始化分级管理员数据
-
-        1. 遍历各个需要初始化的系统
-        2. 查询系统的常用操作与系统的操作信息, 拼装出授权范围
-        3. 返回role info
-        """
-        role_info = RoleInfoBean(
-            name=data["biz_name"],
-            description="管理员可授予他人{}业务的权限".format(data["biz_name"]),
-            members=[ADMIN_USER],
-            subject_scopes=[Subject(type="*", id="*")],
-            authorization_scopes=[],
-        )
-
-        # 默认需要初始化的系统列表
-        systems = ["bk_job", "bk_cmdb", "bk_monitorv3", "bk_log_search", "bk_sops", "bk_nodeman", "bk_gsekit"]
-        for system_id in systems:
-            resource_type = "biz" if system_id != "bk_sops" else "project"
-            resource_system = "bk_cmdb" if system_id != "bk_sops" else "bk_sops"
-            resource_id = data["biz_id"] if system_id != "bk_sops" else data["project_id"]
-            resource_name = data["biz_name"] if system_id != "bk_sops" else data["project_name"]
-
-            auth_scope = AuthScopeSystem(system_id=system_id, actions=[])
-
-            # 1. 查询常用操作
-            common_action = self.role_biz.get_common_action_by_name(
-                system_id, ManagementCommonActionNameEnum.OPS.value
-            )
-            if not common_action:
-                continue
-
-            # 2. 查询操作信息
-            action_list = self.action_biz.list(system_id)
-
-            # 3. 生成授权范围
-            for action_id in common_action.action_ids:
-                action = action_list.get(action_id)
-                if not action:
-                    continue
-
-                # 不关联资源类型的操作
-                if len(action.related_resource_types) == 0:
-                    auth_scope_action = AuthScopeAction(id=action.id, resource_groups=ResourceGroupList(__root__=[]))
-                else:
-                    policy_data = {
-                        "id": action.id,
-                        "resource_groups": [
-                            {
-                                "related_resource_types": [
-                                    {
-                                        "system_id": rrt.system_id,
-                                        "type": rrt.id,
-                                        "condition": [
-                                            {
-                                                "id": gen_uuid(),
-                                                "instances": [
-                                                    {
-                                                        "type": resource_type,
-                                                        "path": [
-                                                            {
-                                                                "id": resource_id,
-                                                                "name": resource_name,
-                                                                "system_id": resource_system,
-                                                                "type": resource_type,
-                                                            }
-                                                        ],
-                                                    }
-                                                ],
-                                                "attributes": [],
-                                            }
-                                        ],
-                                    }
-                                    for rrt in action.related_resource_types
-                                ]
-                            }
-                        ],
-                    }
-                    auth_scope_action = AuthScopeAction.parse_obj(policy_data)
-
-                auth_scope.actions.append(auth_scope_action)
-
-            # 4. 组合授权范围
-            if auth_scope.actions:
-                role_info.authorization_scopes.append(auth_scope)
-
-        return role_info
-
-    def init_group_auth_info(self, authorization_scopes, name_suffix: str):
-        templates = []
-        for auth_scope in authorization_scopes:
-            system_id = auth_scope["system_id"]
-            actions = auth_scope["actions"]
-            if name_suffix == ManagementGroupNameSuffixEnum.READ.value:
-                common_action = self.role_biz.get_common_action_by_name(
-                    system_id, ManagementCommonActionNameEnum.READ.value
-                )
-                if not common_action:
-                    continue
-
-                actions = [a for a in actions if a["id"] in common_action.action_ids]
-
-            policies = [PolicyBean.parse_obj(action) for action in actions]
-            policy_list = PolicyBeanList(
-                system_id=system_id,
-                policies=policies,
-                need_fill_empty_fields=True,  # 填充相关字段
-            )
-
-            template = GroupTemplateGrantBean(
-                system_id=system_id,
-                template_id=0,  # 自定义权限template_id为0
-                policies=policy_list.policies,
-            )
-
-            templates.append(template)
-
-        return templates
